added roles and refactoring

+++ Medicine.BLL/Servises/MedicamentService.cs
@@ -24,12 +24,12 @@ namespace Medicine.BLL.Services
         }
         public void Create(MedicamentDTO item)
         {
-            Database.Medicaments.Create(new Medicament() { Id = (int)item.Id, Name = item.Name });
+            Database.Medicaments.Create(new Medicament() { Id = item.Id, Name = item.Name });
             Database.Save();
         }
         public void Update(MedicamentDTO item)
         {
-            Database.Medicaments.Update(new Medicament() { Id = (int)item.Id, Name = item.Name });
+            Database.Medicaments.Update(new Medicament() { Id = item.Id, Name = item.Name });
             Database.Save();
         }
        public void Delete(int id)
